model/maskrcnn/maskrcnn_requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
av==9.2.0
black==22.3.0
boto3==1.24.54
botocore==1.27.54
cachetools==5.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.5
cycler==0.11.0
Cython==0.29.30
fairscale==0.4.6
fonttools==4.34.4
future==0.18.2
fvcore @ git+https://github.com/facebookresearch/fvcore@15171b62a2ee8e84546af8865f32264739234899
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
iopath==0.1.9
jmespath==1.0.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mypy-extensions==0.4.3
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.2.2
opencv-python==4.6.0.66
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
portalocker==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32==302
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scipy @ file:///C:/ci/scipy_1641555141383/work
six==1.16.0
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
timm==0.6.5
toml==0.10.2
tomli==2.0.1
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.28.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.10
webcolors==1.12
Werkzeug==2.2.0
wincertstore==0.2
yacs==0.1.8
zipp==3.8.1
